refactor(#67): 상영 조회 컨트롤러 변경
