+++ src/main/java/cinebox/service/MovieServiceImpl.java
@@ -1,16 +1,12 @@
 package cinebox.service;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import cinebox.common.enums.MovieStatus;
 import cinebox.common.exception.movie.DuplicatedMovieException;
@@ -18,18 +14,14 @@ import cinebox.common.exception.movie.MovieDeleteFailedException;
 import cinebox.common.exception.movie.NotFoundMovieException;
 import cinebox.dto.request.MovieRequest;
 import cinebox.dto.response.MovieResponse;
-import cinebox.dto.response.ScreenResponse;
 import cinebox.entity.Movie;
-import cinebox.entity.Screen;
 import cinebox.repository.MovieRepository;
-import cinebox.repository.ScreenRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class MovieServiceImpl implements MovieService {
 	private final MovieRepository movieRepository;
-	private final ScreenRepository screenRepository;
 	
 	// 영화 등록(생성)
 	@Override
@@ -80,36 +72,6 @@ public class MovieServiceImpl implements MovieService {
 			throw NotFoundMovieException.EXCEPTION;
 		}
 		return MovieResponse.from(movie);
-	}
-	
-	// 특정 영화 상영 날짜 목록 조회
-	@Override
-	public List<LocalDate> getAvailableDatesForMovie(Long movieId) {
-		Movie movie = movieRepository.findById(movieId)
-				.orElseThrow(() -> NotFoundMovieException.EXCEPTION);
-		
-		List<Screen> screens = movie.getScreens();
-		
-		Set<LocalDate> dateSet = screens.stream()
-				.map(screen -> screen.getStartTime().toLocalDate())
-				.collect(Collectors.toSet());
-		
-		return dateSet.stream().sorted().collect(Collectors.toList());
-	}
-
-	// 날짜별 상영 정보 조회
-	@Override
-	@Transactional(readOnly = true)
-	public List<ScreenResponse> getScreensByDate(Long movieId, LocalDate date) {
-		movieRepository.findById(movieId).orElseThrow(() -> NotFoundMovieException.EXCEPTION);
-		
-		LocalDateTime startOfDay = date.atStartOfDay();
-		LocalDateTime endOfDay = date.plusDays(1).atStartOfDay();
-		
-		List<Screen> screens = screenRepository.findByMovie_MovieIdAndStartTimeBetweenOrderByStartTimeAsc(movieId, startOfDay, endOfDay);
-		return screens.stream()
-				.map(ScreenResponse::from)
-				.collect(Collectors.toList());
 	}
 	
 	// 영화 정보 수정
