+++ src/main/java/cinebox/service/ScreenService.java
@@ -1,5 +1,8 @@
 package cinebox.service;
 
+import java.time.LocalDate;
+import java.util.List;
+
 import cinebox.dto.request.ScreenRequest;
 import cinebox.dto.response.ScreenResponse;
 
@@ -12,4 +15,10 @@ public interface ScreenService {
 
 	// 상영 정보 삭제
 	void deleteScreen(Long screenId);
+
+	// 특정 영화 상영 날짜 목록 조회
+	List<LocalDate> getAvailableDatesForMovie(Long movieId);
+	
+    // 특정 영화의 날짜별 상영 정보 조회
+	List<ScreenResponse> getScreensByDate(Long movieId, LocalDate date);
 }
