+++ src/main/java/cinebox/service/ScreenServiceImpl.java
@@ -1,7 +1,10 @@
 package cinebox.service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -37,11 +40,12 @@ public class ScreenServiceImpl implements ScreenService {
 		Auditorium auditorium = auditoriumRepository.findById(request.auditoriumId())
 				.orElseThrow(() -> NotFoundAuditoriumException.EXCEPTION);
 
-		LocalDateTime endTime = request.startTime().plusMinutes(movie.getRunTime() + 10); // 🎯 endTime 자동 계산
-		
+		LocalDateTime endTime = request.startTime().plusMinutes(movie.getRunTime() + 10);
+
 		List<Screen> overlappingScreens = screenRepository
 				.findByAuditoriumAndStartTimeLessThanAndEndTimeGreaterThan(auditorium, endTime, request.startTime());
-		if(!overlappingScreens.isEmpty()) {
+
+		if (!overlappingScreens.isEmpty()) {
 			throw ScreenTimeConflictException.EXCEPTION;
 		}
 
@@ -61,36 +65,72 @@ public class ScreenServiceImpl implements ScreenService {
 	@Override
 	@Transactional
 	public ScreenResponse updateScreen(Long screenId, ScreenRequest request) {
-		Screen screen = screenRepository.findById(screenId).orElseThrow(() -> NotFoundScreenException.EXCEPTION);
+		Screen screen = screenRepository.findById(screenId)
+				.orElseThrow(() -> NotFoundScreenException.EXCEPTION);
 
 		Movie movie = request.movieId() != null
 				? movieRepository.findById(request.movieId())
 						.orElseThrow(() -> NotFoundMovieException.EXCEPTION)
 				: screen.getMovie();
-		
-		Auditorium auditorium = request.auditoriumId() != null 
-	            ? auditoriumRepository.findById(request.auditoriumId())
+
+		Auditorium auditorium = request.auditoriumId() != null
+				? auditoriumRepository.findById(request.auditoriumId())
 						.orElseThrow(() -> NotFoundAuditoriumException.EXCEPTION)
 				: screen.getAuditorium();
-		
+
 		LocalDateTime startTime = request.startTime() != null ? request.startTime() : screen.getStartTime();
 		LocalDateTime endTime = startTime.plusMinutes(movie.getRunTime() + 10);
-		
+
 		List<Screen> overlappingScreens = screenRepository
 				.findByAuditoriumAndScreenIdNotAndStartTimeLessThanAndEndTimeGreaterThan(auditorium, screenId, endTime, startTime);
-        if (!overlappingScreens.isEmpty()) {
-            throw ScreenTimeConflictException.EXCEPTION;
-        }
+		if (!overlappingScreens.isEmpty()) {
+			throw ScreenTimeConflictException.EXCEPTION;
+		}
 
 		screen.updateScreen(movie, auditorium, startTime, request.price());
-		return ScreenResponse.from(screenRepository.save(screen));
+		Screen saved = screenRepository.save(screen);
+		return ScreenResponse.from(saved);
 	}
 
 	// 상영 정보 삭제
 	@Override
 	@Transactional
 	public void deleteScreen(Long screenId) {
-		Screen screen = screenRepository.findById(screenId).orElseThrow(() -> NotFoundScreenException.EXCEPTION);
+		Screen screen = screenRepository.findById(screenId)
+				.orElseThrow(() -> NotFoundScreenException.EXCEPTION);
 		screenRepository.delete(screen);
+	}
+
+	// 특정 영화 상영 날짜 목록 조회
+	@Override
+	@Transactional(readOnly = true)
+	public List<LocalDate> getAvailableDatesForMovie(Long movieId) {
+		Movie movie = movieRepository.findById(movieId)
+				.orElseThrow(() -> NotFoundMovieException.EXCEPTION);
+
+		List<Screen> screens = movie.getScreens();
+
+		Set<LocalDate> dateSet = screens.stream()
+				.map(screen -> screen.getStartTime().toLocalDate())
+				.collect(Collectors.toSet());
+
+		return dateSet.stream().sorted().collect(Collectors.toList());
+	}
+
+	// 날짜별 상영 정보 조회
+	@Override
+	@Transactional(readOnly = true)
+	public List<ScreenResponse> getScreensByDate(Long movieId, LocalDate date) {
+		movieRepository.findById(movieId).orElseThrow(() -> NotFoundMovieException.EXCEPTION);
+
+		LocalDateTime startOfDay = date.atStartOfDay();
+		LocalDateTime endOfDay = date.plusDays(1).atStartOfDay();
+
+		List<Screen> screens = screenRepository
+				.findByMovie_MovieIdAndStartTimeBetweenOrderByStartTimeAsc(movieId, startOfDay, endOfDay);
+
+		return screens.stream()
+				.map(ScreenResponse::from)
+				.collect(Collectors.toList());
 	}
 }
